+++ component/include/pl_modbus_base.h
@@ -13,20 +13,20 @@ namespace PL {
 class ModbusBase : public virtual Lockable {
 public:
   /// @brief Default transaction buffer size
-  static const size_t defaultBufferSize = 260;
+  static constexpr size_t defaultBufferSize = 260;
   /// @brief Default protocol for network interface
-  static const ModbusProtocol defaultNetworkProtocol = ModbusProtocol::tcp;
+  static constexpr ModbusProtocol defaultNetworkProtocol = ModbusProtocol::tcp;
   /// @brief Default station address for network interface
-  static const uint8_t defaultNetworkStationAddress = 255;
+  static constexpr uint8_t defaultNetworkStationAddress = 255;
 
   /// @brief Maximum number of coils or discrete inputs that can be read in one request
-  static const uint16_t maxNumberOfModbusBitsToRead = 2000;
+  static constexpr uint16_t maxNumberOfModbusBitsToRead = 2000;
   /// @brief Maximum number of coils that can be written in one request
-  static const uint16_t maxNumberOfModbusBitsToWrite = 1968;
+  static constexpr uint16_t maxNumberOfModbusBitsToWrite = 1968;
   /// @brief Maximum number of holding or input registers that can be read in one request
-  static const uint16_t maxNumberOfModbusRegistersToRead = 125;
+  static constexpr uint16_t maxNumberOfModbusRegistersToRead = 125;
   /// @brief Maximum number of holding registers that can be written in one request
-  static const uint16_t maxNumberOfModbusRegistersToWrite = 123;
+  static constexpr uint16_t maxNumberOfModbusRegistersToWrite = 123;
 
   /// @brief Gets Modbus protocol
   /// @return protocol