+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Lock timeout handling.
+- Static const members to constexpr.
 
 ## [1.2.1] - 2024-09-12
 ### Fixed